+++ packages/cad-simple-viewer/src/app/index.ts
@@ -1,4 +1,5 @@
 export * from './AcApContext'
 export * from './AcApDocument'
 export * from './AcApDocManager'
+export * from './AcApGlobalFunc'
 export * from './AcApSettingManager'